target-riscv: sync up codebase with last set of patches submitted upstream for review: https://github.com/sagark/riscv-qemu-upstream2
