+++ disas.c
@@ -312,8 +312,6 @@ void disas(FILE *out, void *code, unsigned long size)
     print_insn = print_insn_s390;
 #elif defined(__ia64__)
     print_insn = print_insn_ia64;
-#elif defined(__riscv__)
-    print_insn = print_insn_riscv;
 #endif
     if (print_insn == NULL) {
         print_insn = print_insn_od_host;
