+++ hw/riscv/Makefile.objs
@@ -1,4 +1,5 @@
-obj-y += riscv_rtc.o riscv_int.o
+obj-y += riscv_rtc.o
+obj-y += riscv_int.o
 obj-y += htif/elf_symb.o
 obj-y += htif/htif.o
 obj-y += riscv_board.o
