+++ target-riscv/Makefile.objs
@@ -1,1 +1,1 @@
-obj-y += translate.o op_helper.o helper.o cpu.o fpu_helper.o gdbstub.o
+obj-y += translate.o op_helper.o helper.o cpu.o fpu_helper.o
