+++ target-riscv/cpu.c
@@ -70,11 +70,6 @@ static void riscv_cpu_reset(CPUState *s)
     cs->exception_index = EXCP_NONE;
 }
 
-static void riscv_cpu_disas_set_info(CPUState *s, disassemble_info *info)
-{
-    info->print_insn = print_insn_riscv;
-}
-
 static void riscv_cpu_realizefn(DeviceState *dev, Error **errp)
 {
     CPUState *cs = CPU(dev);
@@ -123,8 +118,6 @@ static void riscv_cpu_class_init(ObjectClass *c, void *data)
     cc->dump_state = riscv_cpu_dump_state;
     cc->set_pc = riscv_cpu_set_pc;
     cc->synchronize_from_tb = riscv_cpu_synchronize_from_tb;
-    cc->gdb_read_register = riscv_cpu_gdb_read_register;
-    cc->gdb_write_register = riscv_cpu_gdb_write_register;
 #ifdef CONFIG_USER_ONLY
     cc->handle_mmu_fault = riscv_cpu_handle_mmu_fault;
 #else
@@ -134,9 +127,6 @@ static void riscv_cpu_class_init(ObjectClass *c, void *data)
 #endif
     /* For now, mark unmigratable: */
     cc->vmsd = &vmstate_riscv_cpu;
-    cc->disas_set_info = riscv_cpu_disas_set_info;
-    cc->gdb_num_core_regs = 132;
-    cc->gdb_stop_before_watchpoint = true;
 
     /*
      * Reason: riscv_cpu_initfn() calls cpu_exec_init(), which saves
